Move declaration & definition of traceback_callstack into Kokkos_Abort.cpp
