+++ core/src/impl/Kokkos_Error.hpp
@@ -28,8 +28,6 @@ namespace Impl {
 
 [[noreturn]] void throw_runtime_exception(const std::string &msg);
 
-void traceback_callstack(std::ostream &);
-
 std::string human_memory_size(size_t arg_bytes);
 
 }  // namespace Impl
